Auto-detect VS2019 platform
The CMake Visual Studio 2019 Generator does not support the Win64
or ARM suffix on the generator name. Instead the generator platform
must be set explicitly via the platform parameter.

Before this patch the platform was not set at all which results
in CMake picking the host architecture. This obviously leads
to mismatches if a 32bit NodeJS is used on a 64bit Windows.

Right now the platform parameter is only applied for Visual
Studio 16 because cmake-js does not specify a minimum CMake
version. If at least CMake 3.1 (released December 2014)
would be required the platform selection code could be made
unconditional and thus forward-compatible with newer CMake
Visual Studio generators.

Thanks to @WopsS for #177 which added platform parameter support.