Correctly handle spaces in paths
Previously, cmake-js would fail if the cmake path contained spaces (See
issue #167). It would call exec() with a path that was not escaped or quoted,
and fail to run. If we fix that, then spawn fails for similar reasons. Update
both functions to just take the command line as an array of strings.