always remove `Path` if `PATH` is also defined
Bash on Windows defines both PATH and Path. This leads to a crash in MSBuild. Currently, `Path` is only removed from the environment, if it is the not the same as `PATH`. This commit changes this such, that `Path` is always removed, but if the two are not the same, it is appenden to `PATH`, too.